Tryiong to get Grafana working

+++ roles/voltdb/files/setup_graf-prom.sh
@@ -26,6 +26,7 @@ cd "$(dirname "${BASH_SOURCE[0]}")"
 src_dashboard_dir=${HOME}/bin/dashboards
 tgt_dashboard_dir=/etc/grafana/provisioning/dashboards
 tgt_data_dir=/etc/grafana/provisioning/datasources
+PASSWORD="n0mad1c"
 
 # Need these for .vdbhosts and .clusterid
 eth0IP=$(ip -4 -o addr show dev eth0| awk '{split($4,a,"/");print a[1]}')
@@ -62,12 +63,14 @@ function setup_grafana () {
 	sudo cp prometheus_datasource.yaml $tgt_data_dir
 	sudo chgrp grafana ${tgt_dashboard_dir}/grafana_dashboard_signpost.yaml
 	sudo chgrp grafana ${tgt_data_dir}/prometheus_datasource.yaml
-	
+
 	# Start/Enable the service
 	sudo /bin/systemctl daemon-reload
 	sudo /bin/systemctl enable grafana-server
 	sudo /bin/systemctl start grafana-server
 
+	grafana-cli admin reset-admin-password $PASSWORD
+	
 	###---- End Grafana
 }
 
@@ -81,7 +84,16 @@ function setup_prometheus () {
 	bash prometheusserver_configure.sh
 
 	## Disable node_exporter don't need it on mgmt server.
-	sudo systemctl stop prometheus-node-exporter
+	sudo systemctl is-active --quiet prometheus-node-exporter
+
+	if  [ "$?" = "0" ] ; then
+  	    echo stopping node-exporter ...
+	  	sudo systemctl stop prometheus-node-exporter
+	else
+		echo "node-exporter not running, nothing to do"
+	fi
+	
+	# Make sure it doesn't start
 	sudo systemctl disable prometheus-node-exporter
 	
 	###---- End Prometheus
